+++ robozonky-api/src/main/java/com/github/robozonky/internal/util/json/CharArrayAdapter.java
@@ -14,23 +14,19 @@
  * limitations under the License.
  */
 
-package com.github.robozonky.internal.remote.adapters;
+package com.github.robozonky.internal.util.json;
 
 import javax.json.bind.adapter.JsonbAdapter;
 
-import com.github.robozonky.api.Money;
-
-public final class MoneyAdapter implements JsonbAdapter<Money, String> {
+public final class CharArrayAdapter implements JsonbAdapter<char[], String> {
 
     @Override
-    public String adaptToJson(Money obj) {
-        return obj.getValue()
-            .stripTrailingZeros()
-            .toPlainString();
+    public String adaptToJson(char[] obj) {
+        return String.valueOf(obj);
     }
 
     @Override
-    public Money adaptFromJson(String obj) {
-        return Money.from(obj);
+    public char[] adaptFromJson(String obj) {
+        return obj.toCharArray();
     }
 }
