+++ robozonky-api/src/main/java/com/github/robozonky/internal/util/json/PurposeDeserializer.java
@@ -14,20 +14,13 @@
  * limitations under the License.
  */
 
-package com.github.robozonky.api.remote.enums;
+package com.github.robozonky.internal.util.json;
 
-import javax.json.bind.annotation.JsonbTypeDeserializer;
+import com.github.robozonky.api.remote.enums.Purpose;
 
-import com.github.robozonky.internal.util.json.LoanHealthInfoDeserializer;
+public final class PurposeDeserializer extends AbstractDeserializer<Purpose> {
 
-/**
- * {@link #UNKNOWN} is an internal value, not in the Zonky API, and therefore should only ever go last.
- */
-@JsonbTypeDeserializer(LoanHealthInfoDeserializer.class)
-public enum LoanHealth {
-
-    HEALTHY,
-    CURRENTLY_IN_DUE,
-    HISTORICALLY_IN_DUE,
-    UNKNOWN;
+    public PurposeDeserializer() {
+        super(Purpose::valueOf, Purpose.OTHER);
+    }
 }
